Adicionando funcionalidades de autenticação no repositório derivado da main
